labs/S3/resizeImages/lambda/transform.py
import boto3
import os
import sys
import uuid
from urllib.parse import unquote_plus
from PIL import Image
import PIL.Image
            
s3_client = boto3.client('s3')
            
def resize_image(image_path, resized_path):
  with Image.open(image_path) as image:
    image.thumbnail(tuple(x / 2 for x in image.size))
    image.save(resized_path)

def get_record_values(record):
    bucket = record['s3']['bucket']['name']
    key = unquote_plus(record['s3']['object']['key'])
    file_name = key.split("/")[-1]
    return bucket, key, file_name

def get_paths(key):
    tmpkey = key.replace('/', '')
    download_path = '/tmp/{}{}'.format(uuid.uuid4(), tmpkey)
    upload_path = '/tmp/resized-{}'.format(tmpkey)
    return download_path, upload_path

def handler(event, context):
  for record in event['Records']:
    bucket, key, file_name = get_record_values(record)
    download_path, upload_path = get_paths(key)

    s3_client.download_file(bucket, key, download_path)
    resize_image(download_path, upload_path)
    s3_client.upload_file(upload_path, bucket, 'transformed/{}'.format(file_name))
