labs/S3/resizeImages/lambda/requirements.txt
botocore==1.34.41
boto3
pillow
